+++ src/main/java/com/dnd/accompany/domain/accompany/service/AccompanyBoardService.java
@@ -1,7 +1,6 @@
 package com.dnd.accompany.domain.accompany.service;
 
 import static com.dnd.accompany.domain.accompany.entity.AccompanyBoard.*;
-import static com.dnd.accompany.domain.accompany.entity.enums.Role.*;
 
 import java.util.List;
 
@@ -10,24 +9,16 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.dnd.accompany.domain.accompany.api.dto.AccompanyBoardDetailInfo;
 import com.dnd.accompany.domain.accompany.api.dto.AccompanyBoardThumbnail;
 import com.dnd.accompany.domain.accompany.api.dto.CreateAccompanyBoardRequest;
-import com.dnd.accompany.domain.accompany.api.dto.CreateAccompanyBoardResponse;
 import com.dnd.accompany.domain.accompany.api.dto.FindBoardThumbnailResult;
 import com.dnd.accompany.domain.accompany.api.dto.FindBoardThumbnailsResult;
 import com.dnd.accompany.domain.accompany.api.dto.PageResponse;
-import com.dnd.accompany.domain.accompany.api.dto.ReadAccompanyBoardResponse;
-import com.dnd.accompany.domain.accompany.api.dto.UserProfileThumbnail;
 import com.dnd.accompany.domain.accompany.entity.AccompanyBoard;
-import com.dnd.accompany.domain.accompany.entity.AccompanyImage;
 import com.dnd.accompany.domain.accompany.entity.enums.Region;
-import com.dnd.accompany.domain.accompany.exception.accompanyboard.AccompanyBoardAccessDeniedException;
 import com.dnd.accompany.domain.accompany.exception.accompanyboard.AccompanyBoardNotFoundException;
 import com.dnd.accompany.domain.accompany.infrastructure.AccompanyBoardRepository;
-import com.dnd.accompany.domain.user.entity.User;
 import com.dnd.accompany.domain.user.exception.UserNotFoundException;
-import com.dnd.accompany.domain.user.exception.UserProfileNotFoundException;
 import com.dnd.accompany.global.common.response.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
@@ -37,14 +28,10 @@ import lombok.RequiredArgsConstructor;
 public class AccompanyBoardService {
 
 	private final AccompanyBoardRepository accompanyBoardRepository;
-	private final AccompanyImageService accompanyImageService;
-	private final AccompanyTagService accompanyTagService;
-	private final AccompanyUserService accompanyUserService;
-	private final AccompanyRequestService accompanyRequestService;
 
 	@Transactional
-	public CreateAccompanyBoardResponse create(Long userId, CreateAccompanyBoardRequest request) {
-		AccompanyBoard accompanyBoard = accompanyBoardRepository.save(
+	public AccompanyBoard save(CreateAccompanyBoardRequest request) {
+		return accompanyBoardRepository.save(
 			builder()
 				.title(request.title())
 				.content(request.content())
@@ -58,18 +45,13 @@ public class AccompanyBoardService {
 				.preferredGender(request.preferredGender())
 				.build()
 		);
-		accompanyImageService.save(accompanyBoard, request.imageUrls());
-		accompanyTagService.save(accompanyBoard, request.tagNames());
-		accompanyUserService.save(userId, accompanyBoard, HOST);
-
-		return new CreateAccompanyBoardResponse(accompanyBoard.getId());
 	}
 
 	@Transactional(readOnly = true)
-	public PageResponse<AccompanyBoardThumbnail> readAll(Pageable pageable, Region region) {
+	public PageResponse<AccompanyBoardThumbnail> getAllBoards(Pageable pageable, Region region) {
 		Slice<FindBoardThumbnailsResult> sliceResult = accompanyBoardRepository.findBoardThumbnails(pageable, region);
 
-		List<AccompanyBoardThumbnail> thumbnails = getAccompanyBoardThumbnails(sliceResult.getContent());
+		List<AccompanyBoardThumbnail> thumbnails = getBoardThumbnails(sliceResult.getContent());
 
 		return new PageResponse<>(sliceResult.hasNext(), thumbnails);
 	}
@@ -77,7 +59,7 @@ public class AccompanyBoardService {
 	/**
 	 * imageUrls의 타입을 String -> List<String>로 변환합니다.
 	 */
-	private static List<AccompanyBoardThumbnail> getAccompanyBoardThumbnails(List<FindBoardThumbnailsResult> results) {
+	private static List<AccompanyBoardThumbnail> getBoardThumbnails(List<FindBoardThumbnailsResult> results) {
 		List<AccompanyBoardThumbnail> thumbnails = results.stream()
 			.map(result -> AccompanyBoardThumbnail.builder()
 				.boardId(result.boardId())
@@ -92,66 +74,26 @@ public class AccompanyBoardService {
 		return thumbnails;
 	}
 
-	@Transactional(readOnly = true)
-	public ReadAccompanyBoardResponse read(Long boardId) {
-		AccompanyBoardDetailInfo boardDetailInfo = getAccompanyBoardDetailInfo(boardId);
-		UserProfileThumbnail profileThumbnail = getUserProfileThumbnail(boardId);
-
-		return new ReadAccompanyBoardResponse(boardDetailInfo, profileThumbnail);
-	}
-
-	private UserProfileThumbnail getUserProfileThumbnail(Long boardId) {
-		Long userId = accompanyUserService.findUserIdByAccompanyBoardId(boardId)
-			.orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND));
-
-		UserProfileThumbnail profileThumbnail = accompanyBoardRepository.findUserProfileThumbnail(userId)
-			.orElseThrow(() -> new UserProfileNotFoundException(ErrorCode.PROFILE_NOT_FOUND));
-		return profileThumbnail;
-	}
-
-	private AccompanyBoardDetailInfo getAccompanyBoardDetailInfo(Long boardId) {
-		AccompanyBoard accompanyBoard = accompanyBoardRepository.findByIdWithCategories(boardId)
+	public AccompanyBoard getAccompanyBoard(Long boardId){
+		return accompanyBoardRepository.findByIdWithCategories(boardId)
 			.orElseThrow(() -> new AccompanyBoardNotFoundException(ErrorCode.ACCOMPANY_BOARD_NOT_FOUND));
-
-		List<String> tagNames = accompanyTagService.findTagNamesByAccompanyBoardId(boardId);
-		List<String> imageUrls = accompanyImageService.findImageUrlsByAccompanyBoardId(boardId);
-
-		AccompanyBoardDetailInfo boardDetailInfo = new AccompanyBoardDetailInfo(accompanyBoard, tagNames, imageUrls);
-		return boardDetailInfo;
 	}
 
-	@Transactional
-	public void delete(Long userId, Long boardId) {
-		if (accompanyBoardRepository.isHostOfBoard(userId, boardId)) {
-			accompanyImageService.deleteByBoardId(boardId);
-			accompanyTagService.deleteByBoardId(boardId);
-			accompanyRequestService.deleteByBoardId(boardId);
-			accompanyUserService.deleteByBoardId(boardId);
-			accompanyBoardRepository.deleteById(boardId);
-		} else {
-			throw new AccompanyBoardAccessDeniedException(ErrorCode.ACCOMPANY_BOARD_ACCESS_DENIED);
-		}
+	public boolean isHostOfBoard(Long userId, Long boardId){
+		return accompanyBoardRepository.isHostOfBoard(userId, boardId);
 	}
 
-	public AccompanyBoardThumbnail findAccompanyThumbnail(Long boardId, Long userId){
-		FindBoardThumbnailResult result = accompanyBoardRepository.findBoardThumbnail(boardId)
+	public void deleteByBoardId(Long boardId){
+		accompanyBoardRepository.deleteById(boardId);
+	}
+
+	public FindBoardThumbnailResult getBoardThumbnail(Long boardId){
+		return accompanyBoardRepository.findBoardThumbnail(boardId)
 			.orElseThrow(() -> new AccompanyBoardNotFoundException(ErrorCode.ACCOMPANY_BOARD_NOT_FOUND));
+	}
 
-		String nickname = accompanyBoardRepository.findNickname(userId)
+	public String getNickname(Long userId){
+		return accompanyBoardRepository.findNickname(userId)
 			.orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND));
-
-		List<String> imageUrls = accompanyImageService.findImageUrlsByAccompanyBoardId(boardId);
-
-		AccompanyBoardThumbnail boardThumbnail = AccompanyBoardThumbnail.builder()
-			.boardId(result.boardId())
-			.title(result.title())
-			.region(result.region())
-			.startDate(result.startDate())
-			.endDate(result.endDate())
-			.nickname(nickname)
-			.imageUrls(imageUrls)
-			.build();
-
-		return boardThumbnail;
 	}
 }
