+++ src/main/java/com/dnd/accompany/domain/accompany/service/AccompanyServiceFacade.java
@@ -1,0 +1,128 @@
+package com.dnd.accompany.domain.accompany.service;
+
+import static com.dnd.accompany.domain.accompany.entity.enums.Role.*;
+
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.dnd.accompany.domain.accompany.api.dto.AccompanyBoardDetailInfo;
+import com.dnd.accompany.domain.accompany.api.dto.AccompanyBoardThumbnail;
+import com.dnd.accompany.domain.accompany.api.dto.AccompanyRequestDetailInfo;
+import com.dnd.accompany.domain.accompany.api.dto.CreateAccompanyBoardRequest;
+import com.dnd.accompany.domain.accompany.api.dto.CreateAccompanyBoardResponse;
+import com.dnd.accompany.domain.accompany.api.dto.FindBoardThumbnailResult;
+import com.dnd.accompany.domain.accompany.api.dto.ReadAccompanyBoardResponse;
+import com.dnd.accompany.domain.accompany.api.dto.ReadAccompanyRequest;
+import com.dnd.accompany.domain.accompany.api.dto.ReadAccompanyResponse;
+import com.dnd.accompany.domain.accompany.api.dto.UserProfileDetailInfo;
+import com.dnd.accompany.domain.accompany.api.dto.UserProfileThumbnail;
+import com.dnd.accompany.domain.accompany.entity.AccompanyBoard;
+import com.dnd.accompany.domain.accompany.exception.accompanyboard.AccompanyBoardAccessDeniedException;
+import com.dnd.accompany.domain.user.dto.UserProfileDetailResponse;
+import com.dnd.accompany.domain.user.service.UserProfileService;
+import com.dnd.accompany.global.common.response.ErrorCode;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class AccompanyServiceFacade {
+
+	private final AccompanyBoardService accompanyBoardService;
+	private final AccompanyImageService accompanyImageService;
+	private final AccompanyTagService accompanyTagService;
+	private final AccompanyUserService accompanyUserService;
+	private final AccompanyRequestService accompanyRequestService;
+	private final UserProfileService userProfileService;
+
+	@Transactional
+	public CreateAccompanyBoardResponse createBoard(Long userId, CreateAccompanyBoardRequest request) {
+		AccompanyBoard accompanyBoard = accompanyBoardService.save(request);
+
+		accompanyImageService.save(accompanyBoard, request.imageUrls());
+		accompanyTagService.save(accompanyBoard, request.tagNames());
+		accompanyUserService.save(userId, accompanyBoard, HOST);
+
+		return new CreateAccompanyBoardResponse(accompanyBoard.getId());
+	}
+
+	@Transactional(readOnly = true)
+	public ReadAccompanyBoardResponse getBoardDetail(Long boardId) {
+		AccompanyBoardDetailInfo boardDetailInfo = getBoardDetailInfo(boardId);
+		UserProfileThumbnail profileThumbnail = accompanyUserService.getUserProfileThumbnail(boardId);
+
+		return new ReadAccompanyBoardResponse(boardDetailInfo, profileThumbnail);
+	}
+
+	private AccompanyBoardDetailInfo getBoardDetailInfo(Long boardId) {
+		AccompanyBoard accompanyBoard = accompanyBoardService.getAccompanyBoard(boardId);
+
+		List<String> tagNames = accompanyTagService.getTagNames(boardId);
+		List<String> imageUrls = accompanyImageService.getImageUrls(boardId);
+
+		AccompanyBoardDetailInfo boardDetailInfo = new AccompanyBoardDetailInfo(accompanyBoard, tagNames, imageUrls);
+		return boardDetailInfo;
+	}
+
+	@Transactional
+	public void deleteBoard(Long userId, Long boardId) {
+		if (accompanyBoardService.isHostOfBoard(userId, boardId)) {
+			accompanyImageService.deleteByBoardId(boardId);
+			accompanyTagService.deleteByBoardId(boardId);
+			accompanyRequestService.deleteByBoardId(boardId);
+			accompanyUserService.deleteByBoardId(boardId);
+			accompanyBoardService.deleteByBoardId(boardId);
+		} else {
+			throw new AccompanyBoardAccessDeniedException(ErrorCode.ACCOMPANY_BOARD_ACCESS_DENIED);
+		}
+	}
+
+	@Transactional(readOnly = true)
+	public AccompanyBoardThumbnail getBoardThumbnail(Long boardId, Long userId) {
+		FindBoardThumbnailResult result = accompanyBoardService.getBoardThumbnail(boardId);
+		String nickname = accompanyBoardService.getNickname(userId);
+		List<String> imageUrls = accompanyImageService.getImageUrls(boardId);
+
+		AccompanyBoardThumbnail boardThumbnail = AccompanyBoardThumbnail.builder()
+			.boardId(result.boardId())
+			.title(result.title())
+			.region(result.region())
+			.startDate(result.startDate())
+			.endDate(result.endDate())
+			.nickname(nickname)
+			.imageUrls(imageUrls)
+			.build();
+
+		return boardThumbnail;
+	}
+
+	@Transactional(readOnly = true)
+	public ReadAccompanyResponse getRequestDetail(ReadAccompanyRequest request) {
+		AccompanyBoardThumbnail boardThumbnail = getBoardThumbnail(request.boardId(), request.userId());
+
+		UserProfileDetailInfo profileDetailInfo = getUserProfileDetailInfo(request.userId());
+
+		AccompanyRequestDetailInfo requestDetailInfo = accompanyRequestService.getRequestDetailInfo(request.boardId(),
+			request.userId(), request.role());
+
+		return new ReadAccompanyResponse(boardThumbnail, profileDetailInfo, requestDetailInfo);
+	}
+
+	private UserProfileDetailInfo getUserProfileDetailInfo(Long userId) {
+		UserProfileDetailResponse response = userProfileService.findUserProfileDetails(userId);
+
+		return UserProfileDetailInfo.builder()
+			.userId(response.userId())
+			.nickname(response.nickname())
+			.profileImageUrl(response.profileImageUrl())
+			.birthYear(response.birthYear())
+			.gender(response.gender())
+			.travelPreferences(response.travelPreferences())
+			.travelStyles(response.travelStyles())
+			.foodPreferences(response.foodPreferences())
+			.userImageUrls(response.userImageUrls())
+			.build();
+	}
+}
