+++ src/main/java/com/dnd/accompany/domain/accompany/service/AccompanyUserService.java
@@ -5,12 +5,16 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.dnd.accompany.domain.accompany.api.dto.UserProfileThumbnail;
 import com.dnd.accompany.domain.accompany.entity.AccompanyBoard;
 import com.dnd.accompany.domain.accompany.entity.AccompanyUser;
 import com.dnd.accompany.domain.accompany.entity.enums.Role;
 import com.dnd.accompany.domain.accompany.infrastructure.AccompanyBoardRepository;
 import com.dnd.accompany.domain.accompany.infrastructure.AccompanyUserRepository;
 import com.dnd.accompany.domain.user.entity.User;
+import com.dnd.accompany.domain.user.exception.UserNotFoundException;
+import com.dnd.accompany.domain.user.exception.UserProfileNotFoundException;
+import com.dnd.accompany.global.common.response.ErrorCode;
 
 import lombok.RequiredArgsConstructor;
 
@@ -29,8 +33,13 @@ public class AccompanyUserService {
 	}
 
 	@Transactional(readOnly = true)
-	public Optional<Long> findUserIdByAccompanyBoardId(Long boardId) {
-		return accompanyUserRepository.findUserIdByAccompanyBoardId(boardId);
+	public UserProfileThumbnail getUserProfileThumbnail(Long boardId) {
+		Long userId = accompanyUserRepository.findUserIdByAccompanyBoardId(boardId)
+			.orElseThrow(() -> new UserNotFoundException(ErrorCode.USER_NOT_FOUND));
+
+		UserProfileThumbnail profileThumbnail = accompanyUserRepository.findUserProfileThumbnail(userId)
+			.orElseThrow(() -> new UserProfileNotFoundException(ErrorCode.PROFILE_NOT_FOUND));
+		return profileThumbnail;
 	}
 
 	@Transactional
